+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/project/renderer/extension/commonmark/CommonMarkBuildProjectRendererTestExtension.java
@@ -1,0 +1,31 @@
+package net.splitcells.website.server.project.renderer.extension.commonmark;
+
+import net.splitcells.website.server.Config;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+
+import java.nio.file.Path;
+import java.util.Optional;
+
+import static net.splitcells.dem.testing.TestTypes.INTEGRATION_TEST;
+import static net.splitcells.website.server.project.ProjectRenderer.projectRenderer;
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class CommonMarkBuildProjectRendererTestExtension {
+    /**
+     * TODO Test extension explicitly.
+     */
+    @Tag(INTEGRATION_TEST)
+    @Test
+    public void testBuildLayout() {
+        final var testSubject = projectRenderer
+                ("public"
+                        , Path.of("../..")
+                        , Path.of("../net.splitcells.website.default.content/src/main/xsl/net/splitcells/website/den/translation/to/html/")
+                        , Path.of("net.splitcells.website.default.content/src/main/resources/html")
+                        , "/net/splitcells"
+                        , a -> Optional.empty()
+                        , Config.create());
+        assertThat(testSubject.projectPaths()).contains(Path.of("net/splitcells/network/BUILD.html"));
+    }
+}